+++ packages/router-cli/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @tanstack/router-cli
 
+## 1.166.43
+
+### Patch Changes
+
+- Updated dependencies [[`c5811aa`](https://github.com/TanStack/router/commit/c5811aacb59b69a8179f14cd9529a51b2206d504)]:
+  - @tanstack/router-generator@1.166.42
+
+## 1.166.42
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.166.41
+
+## 1.166.41
+
+### Patch Changes
+
+- Updated dependencies [[`056337e`](https://github.com/TanStack/router/commit/056337ef02ccbeaf45ec8533a227761be869e8f3)]:
+  - @tanstack/router-generator@1.166.40
+
+## 1.166.40
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.166.39
+
 ## 1.166.39
 
 ### Patch Changes