+++ packages/router-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tanstack/router-core
 
+## 1.169.2
+
+### Patch Changes
+
+- Update seroval dependencies to version 1.5.4. ([#7340](https://github.com/TanStack/router/pull/7340))
+
+## 1.169.1
+
+### Patch Changes
+
+- Fix params.parse inference for discriminated union path params while preserving path key validation. ([#7306](https://github.com/TanStack/router/pull/7306))
+
 ## 1.169.0
 
 ### Minor Changes